# Conflicts:
#	features/content/play_broadcaster/src/main/java/com/tokopedia/play/broadcaster/data/socket/PlayBroadcastWebSocketMapper.kt
#	features/content/play_broadcaster/src/main/java/com/tokopedia/play/broadcaster/ui/mapper/PlayBroadcastUiMapper.kt
#	features/content/play_broadcaster/src/main/java/com/tokopedia/play/broadcaster/view/viewmodel/PlayBroadcastViewModel.kt
#	features/content/play_broadcaster/src/test/java/com/tokopedia/play/broadcaster/model/interactive/InteractiveUiModelBuilder.kt

+++ features/content/play_common/src/main/java/com/tokopedia/play_common/model/mapper/PlayChannelInteractiveMapper.kt
@@ -14,7 +14,7 @@ import javax.inject.Inject
 class PlayChannelInteractiveMapper @Inject constructor() {
 
     fun mapInteractive(input: GiveawayResponse) = PlayCurrentInteractiveModel(
-        id = input.interactiveID.toLong(),
+        id = input.interactiveID,
         type = InteractiveType.getByValue(input.interactiveType),
         title = input.title,
         timeStatus = PlayInteractiveTimeStatus.getByValue(
